I guess just raw sql then

+++ src/session_store.rs
@@ -9,7 +9,7 @@ use serde::{Serialize, Deserialize};
 
 #[async_trait]
 pub trait SessionStoreFns: Sync + Send {
-    async fn create_session(&self, user_id: i32, profile_id: i32) -> Result<Session, ()>;
+    async fn create_session(&self, user_id: i64, profile_id: i64) -> Result<Session, ()>;
     async fn get_data_of_session(&self, session_id: String) -> Result<SessionValueInStore, ()>;
     async fn get_sessions_of_user(&self);
 }
@@ -22,8 +22,8 @@ pub struct SessionStoreConnection {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct SessionValueInStore {
-    pub user_id: i32,
-    pub profile_id: i32,
+    pub user_id: i64,
+    pub profile_id: i64,
 }
 
 impl SessionStoreConnection {
@@ -40,7 +40,7 @@ impl SessionStoreConnection {
 
 #[async_trait]
 impl SessionStoreFns for SessionStoreConnection {
-    async fn create_session(&self, user_id: i32, profile_id: i32) -> Result<Session, ()> {
+    async fn create_session(&self, user_id: i64, profile_id: i64) -> Result<Session, ()> {
         let session_id = Uuid::new_v4().to_string();
         let session_value_json =
             match serde_json::to_string(&SessionValueInStore {
